feat_get_dm
