+++ src/main/java/com/github/redouane59/twitter/dto/dm/DmAnswer.java
@@ -1,0 +1,14 @@
+package com.github.redouane59.twitter.dto.dm;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class DmAnswer {
+
+  private DirectMessage event;
+
+}
